+++ source/systems/polygongraphics.d
@@ -6,7 +6,7 @@ import std.datetime;
 import std.stdio;
 
 import glamour.texture;
-import gl3n.linalg; 
+import gl3n.linalg;
 
 import camera;
 import components.collider;
@@ -26,41 +26,41 @@ class PolygonGraphics : System!Polygon
     this.xres = xres; this.yres = yres;
     this.camera = camera;
   }
-  
+
   override bool canAddEntity(Entity entity)
   {
-    return "position" in entity.values && "angle" in entity.values && 
+    return "position" in entity.values && "angle" in entity.values &&
            Polygon.canMakeComponent(entity.values);
   }
-  
+
   override Polygon makeComponent(Entity entity)
   {
     Polygon component;
-    
+
     // TODO: maybe split into separate systems for drawing polygons/texts/sprites?
     if ("polygon.vertices" in entity.values)
     {
       if ("polygon.colors" in entity.values)
-        component = new Polygon(entity.values["polygon.vertices"].myTo!(vec2[]), 
+        component = new Polygon(entity.values["polygon.vertices"].myTo!(vec2[]),
                                 entity.values["polygon.colors"].myTo!(vec4[]));
       else if ("color" in entity.values)
-        component = new Polygon(entity.values["polygon.vertices"].myTo!(vec2[]), 
+        component = new Polygon(entity.values["polygon.vertices"].myTo!(vec2[]),
                                 entity.values["color"].myTo!vec4);
     }
-    
+
     component.position = vec2(entity.values["position"].to!(float[2]));
     component.angle = entity.values["angle"].to!double;
-    
+
     return component;
   }
-  
+
   override void updateValues()
   {
     StopWatch debugTimer;
     debugTimer.start;
     vertices = null;
     colors = null;
-    
+
     foreach (component; components)
     {
       auto transform = (vec2 vertex) => ((vec3(vertex, 0.0)*mat3.zrotation(-component.angle)).xy +
@@ -73,22 +73,22 @@ class PolygonGraphics : System!Polygon
     }
     debugText = format("polygongraphics timings: %s", debugTimer.peek.usecs*0.001);
   }
-  
+
   override void updateEntities()
   {
   }
-  
+
   override void updateFromEntities()
   {
-    foreach (int index, Entity entity; entityForIndex)
+    foreach (ulong index, Entity entity; entityForIndex)
     {
       //currentStates[index].velocity = entity.vectors["velocity"];
       components[index].position = vec2(entity.values["position"].to!(float[2]));
       components[index].angle = entity.values["angle"].to!double;
     }
   }
-  
-  
+
+
   immutable int xres, yres;
   Camera camera;
   vec2[][string] vertices;
